mysite/requirements.txt
Django~=4.2.0
djangorestframework
djangorestframework-simplejwt
django-filter
Pillow
django-storages
boto3
psycopg2-binary
gunicorn
whitenoise
pytest
requests
celery
redis
django-environ
django-redis
django-mptt
django-imagekit
